FEAT : 중복 부분으로 인해 상속 프래그먼트로 변경
